[cli] Use `uint16_t` instead of `int`
Per review:
https://github.com/openthread/openthread/pull/4003#discussion_r304727719

+++ src/cli/cli_uart.cpp
@@ -250,7 +250,7 @@ int Uart::Output(const char *aBuf, uint16_t aBufLength)
             sendLength = remaining;
         }
 
-        for (int i = 0; i < sendLength; i++)
+        for (uint16_t i = 0; i < sendLength; i++)
         {
             tail            = (mTxHead + mTxLength) % kTxBufferSize;
             mTxBuffer[tail] = *aBuf++;
